+++ sdks/python/apache_beam/io/iobase.py
@@ -716,7 +716,7 @@ class Write(ptransform.PTransform):
             'sink_dd': self.sink}
 
   def expand(self, pcoll):
-    from apache_beam.runners.dataflow.native_io import iobase as dataflow_io
+    from apache_beam.runners.google_cloud_dataflow.native_io import iobase as dataflow_io
     if isinstance(self.sink, dataflow_io.NativeSink):
       # A native sink
       return pcoll | 'NativeWrite' >> dataflow_io._NativeWrite(self.sink)
@@ -841,4 +841,4 @@ class _RoundRobinKeyFn(core.DoFn):
 
 # For backwards compatibility.
 # pylint: disable=wrong-import-position
-from apache_beam.runners.dataflow.native_io.iobase import *
+from apache_beam.runners.google_cloud_dataflow.native_io.iobase import *
