+++ sdks/python/apache_beam/runners/dataflow_runner.py
@@ -32,7 +32,7 @@ from apache_beam import pvalue
 from apache_beam.internal import json_value
 from apache_beam.internal import pickler
 from apache_beam.pvalue import PCollectionView
-from apache_beam.runners.dataflow.dataflow_metrics import DataflowMetrics
+from apache_beam.runners.google_cloud_dataflow.dataflow_metrics import DataflowMetrics
 from apache_beam.runners.runner import PipelineResult
 from apache_beam.runners.runner import PipelineRunner
 from apache_beam.runners.runner import PipelineState
